Change list content & add cursor /!\ Don't compile
/!\ Don’t compile - commit with unfinished work.

Changes :
 - Use global struct with the list and the cursor in it instead of a
simple list.
‘Cursor’ is the index of the selected item.
 - Debug calls removed.
 - Display a message when the window is too small to display the list.

Add :
 - cursor_go_next : move the cursor to the next item. If the cursor is
on the last item, move it to the first item.
 - cursor_go_prev : move the cursor to the previous item. If the cursor
is on the first item, move it to the last item.
 - cursor_go_left : move the cursor on the left. If the cursor is on
the first column, move it to the same line on the last column.
 - cursor_go_right : move the cursor on the right. If the cursor is on
the last column, move it to the same line on the first column.